+++ app/src/main/kotlin/br/com/rodrigogurgel/catalog/framework/adapter/input/rest/dto/response/common/PageResponseDTO.kt
@@ -3,6 +3,6 @@ package br.com.rodrigogurgel.catalog.framework.adapter.input.rest.dto.response.c
 abstract class PageResponseDTO<T>(
     open val limit: Int,
     open val nextCursor: String?,
-    open val total: Long,
+    open val total: Int,
     open val data: List<T>,
 )
