+++ app/src/main/kotlin/br/com/rodrigogurgel/catalog/framework/adapter/input/rest/dto/response/product/ProductPageResponseDTO.kt
@@ -5,6 +5,6 @@ import br.com.rodrigogurgel.catalog.framework.adapter.input.rest.dto.response.co
 data class ProductPageResponseDTO(
     override val limit: Int,
     override val nextCursor: String?,
-    override val total: Long,
+    override val total: Int,
     override val data: List<ProductResponseDTO>,
 ) : PageResponseDTO<ProductResponseDTO>(limit, nextCursor, total, data)
